Package stacktrace related things into a library
