+++ dep/CMakeLists.txt
@@ -238,8 +238,8 @@ message(STATUS "LLVM headers location: ${LLVM_INCLUDE_DIRS}")
 add_subdirectory(fmt)
 
 # build libbacktrace
-set(LIBBACKTRACE_LIBS ${CMAKE_CURRENT_LIST_DIR}/libbacktrace/.libs/libbacktrace.a)
-set(LIBBACKTRACE_INCLUDE_DIRS ${CMAKE_CURRENT_LIST_DIR}/libbacktrace)
+set(LIBBACKTRACE_LIBS ${CMAKE_CURRENT_LIST_DIR}/libbacktrace/.libs/libbacktrace.a CACHE INTERNAL "libbacktrace libraries")
+set(LIBBACKTRACE_INCLUDE_DIRS ${CMAKE_CURRENT_LIST_DIR}/libbacktrace CACHE INTERNAL "libbacktrace include dirs")
 add_custom_target(libbacktrace ALL DEPENDS ${LIBBACKTRACE_LIBS})
 add_custom_command(
         OUTPUT ${LIBBACKTRACE_LIBS}
@@ -252,14 +252,12 @@ add_custom_command(
 set(DEP_INCLUDES
         ${LLVM_INCLUDE_DIRS}
         ${CLANG_INCLUDE_DIRS}
-        ${LIBBACKTRACE_INCLUDE_DIRS}
         ${DEP_DIR}/filesystem/include
         ${DEP_DIR}/libbacktrace
         PARENT_SCOPE)
 
 set(DEP_LIBS ${LLVM_LIBS}
         ${CLANG_LIBRARIES}
-        ${LIBBACKTRACE_LIBS}
         fmt::fmt
         z dl rt tinfo pthread m xml2
         gcov
@@ -269,5 +267,4 @@ set(DEP_LIBS ${LLVM_LIBS}
 set(DEP_LIB_DIRS
         ${LLVM_LIBRARY_DIR}
         ${CLANG_LIBDIRS}
-        ${DEP_DIR}/libbacktrace/.libs
         PARENT_SCOPE)
