+++ src/backtrace/tan_backtrace.h
@@ -1,0 +1,7 @@
+#ifndef __TAN_SRC_BACKTRACE_TAN_BACKTRACE_H__
+#define __TAN_SRC_BACKTRACE_TAN_BACKTRACE_H__
+
+extern "C" void init_back_trace(const char *filename);
+extern "C" void print_back_trace();
+
+#endif
